+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Inspection: do not use --platform flag with FROM
 - Inspection: wget without recommended flags
 - Inspection: COPY to a relative destination without WORKDIR set
+- Inspection: instruction keywords should use consistent casing
 
 ## [2.0.16] 07-02-2026